Project endpoint

+++ src/main/java/com/worktracker/service/ProjectService.java
@@ -6,5 +6,5 @@ public interface ProjectService {
 
     void createProject(ProjectRequestDTO projectRequestDTO);
 
-    void updateProjectName(Long id, String name);
+    void updateProjectName(Integer id, String name);
 }
